make it possible to add new customers

+++ src/project0/CloseAccountMenu.java
@@ -1,21 +1,22 @@
 package project0;
 
-public class CloseAccountMenu implements IMenu {
+import java.util.Scanner;
+
+public class CloseAccountMenu extends ABCMenu {
 
 	private String accountNumber;
 	private String confirmation;
-	private IMenuTree mt;
 
-	public CloseAccountMenu(IMenuTree mt) {
-		this.mt = mt;
+	public CloseAccountMenu(IMenuTree mt, Scanner s) {
+		super(mt,s);
 	}
 
 	public void Run() {
 		System.out.print("Please enter the account number:\n");
-		this.accountNumber = System.console().readLine();
+		this.accountNumber = s.nextLine();
 		System.out.print("You have entered the account number " + this.accountNumber
 				+ "\nAre you sure you wish to close this account? Y/n");
-		this.confirmation = System.console().readLine();
+		this.confirmation = s.nextLine();
 	}
 
 	private void Execute() {
@@ -41,9 +42,14 @@ public class CloseAccountMenu implements IMenu {
 		try {
 			IDB db = BankDB.getDB();
 			Account a = db.getAccountByNumber(this.accountNumber);
+
 			if (a != null) {
-				db.deleteAccountByNumber(this.accountNumber);
-				System.out.println("Account closed.");
+				if (db.customerOwnsAccount(UserSession.getSession().getID(), this.accountNumber)) {
+					db.deleteAccountByNumber(this.accountNumber);
+					System.out.println("Account closed.");
+				} else {
+					System.out.println("you don\'t own that account");
+				}
 			} else {
 				System.out.println("Account not found.");
 			}
